requirements.txt
tqdm
matplotlib
wandb
hydra-core

numpy
torch
torchvision
torchaudio

